pyproject.toml
[tool.poetry]
name = "djangodrftelegram"
version = "0.1.0"
description = ""
authors = ["kremniland <kremnilandk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
django-cors-headers = "^3.14.0"
django-phonenumber-field = "^7.0.2"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
drf-spectacular = "^0.26.0"
phonenumbers = "^8.13.7"
pillow = "^9.4.0"
psycopg2 = "^2.9.5"
django-environ = "^0.10.0"
loguru = "^0.6.0"
django-crum = "^0.7.9"
aiogram = "^2.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
